[groot-docker-build] use temp dirs
Previously fed PATH=. which causes docker to create a context
out of everything in that directory (e.g. if in your home
dir, this can be large and recursive). Not that it uses
any of that context when you pass it --file, but can
take a considerable amount of time wrapping it up.